genes/human/C18orf21/C18orf21-bioinformatics/pyproject.toml
[project]
name = "c18orf21-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "biopython>=1.85",
    "matplotlib>=3.10.5",
    "numpy>=2.3.2",
    "pandas>=2.3.2",
    "requests>=2.32.5",
    "scikit-learn>=1.7.1",
    "scipy>=1.16.1",
    "seaborn>=0.13.2",
]
